requirements.txt
duckduckgo_search==3.*
Pillow==10.*
fastdownload==0.0.*